+++ examples/custom_gate.rs
@@ -29,7 +29,8 @@ fn main() -> Result<(), QuantrError> {
 
     // Prints the simulation process of each gate (excluding identity gates).
     qc.toggle_simulation_progress();
-    let simulated = qc.simulate();
+    let mut simulated = qc.simulate();
+    simulated.toggle_warnings();
 
     // Prints the bin count of measured states.
     if let Measurement::Observable(bin_count) = simulated.measure_all(50) {